Rework of the Inbox.
In order to make inbox more scalable it had to be reworked from scratch.

Events are stored in ETS table, with unique keys of form:
  {VertexId, EventIdx}
where EventIdx is counter belonging to specific vertex, which can be
atomically incremented (because it's a field in another ETS table).

When rest_edges handler process calls function inbox:post/2 it
atomically increases mentioned counter for proper vertex and places
event in the ETS table. After this, if that vertex isn't worked on
inside processor, a message is sent to Dispatcher process to consider
scheduling event from that vertex's queue.

From the processor perspective, when the work on some event is done, the
'Free' flag of owning vertex is set to 'true' and message is sent to
Dispatcher to consider scheduling another event from that vertex's
queue.

Dispatcher continuously tries to schedule event from as many queues as
possible, and then reads all messages, marking queues that should be
considered for scheduling in the next iteration.