Add jump counter to naive queries
In order to prevent requests being sent indefenitely in case of a cycle
I added jump counter to naive queries.

This counter is incremented each time a request is being sent (or
function called) recursively (its basically size of a call stack).

If jump counter is too big (arbitrary decision: 1000 calls) then
query assumes a cycle was found and returns an error.