Add response status checking inside tests
In order to make sure that operations in the WHEN section are being
executed without errors I added checking if 'client' module calls
return 'ok' atom.

NOTE: there is one operation in naive_queries suite that always fails.
I suppose that's its unintentional mistake and it should not be there
at all, but I've left it and just put a comment in.